requirements.txt
Django==1.8.2
django-debug-toolbar==1.3.2
django-extensions==1.5.5
djangorestframework==3.1.3
gnureadline==6.3.3
ipython==3.2.0
six==1.9.0
sqlparse==0.1.15
Werkzeug==0.10.4
